+++ R/edit.nb.R
@@ -162,6 +162,10 @@ edit.nb <- function(name, coords, polys=NULL, ..., use_region.id=FALSE) {
   if (is.null(icl)) class(nb) <- "nb"
   else class(nb) <- c("nb", icl)
   nb <- sym.attr.nb(nb)
+  if (get.SubgraphOption()) {
+     nsg <- n.comp.nb(nb)$nc
+     if (nsg > 1) warning("neighbour object has ", nsg, " sub-graphs")
+  }
   nb
 }