+++ R/knn2nb.R
@@ -27,5 +27,9 @@ knn2nb <- function(knn, row.names=NULL, sym=FALSE) {
 	attr(res, "type") <- "knn"
  	attr(res, "knn-k") <- knn$k
 	class(res) <- "nb"
+        if (get.SubgraphOption()) {
+          nsg <- n.comp.nb(res)$nc
+          if (nsg > 1) warning("neighbour object has ", nsg, " sub-graphs")
+        }
 	res
 }