+++ R/nboperations.R
@@ -28,6 +28,10 @@ union.nb<-function(nb.obj1, nb.obj2){
   attr(new.nb,"type")<-paste("union(",attr(nb.obj1,"type"),
                              ",",attr(nb.obj2,"type"),")")
   class(new.nb)<-"nb"
+  if (get.SubgraphOption()) {
+    nsg <- n.comp.nb(new.nb)$nc
+    if (nsg > 1) warning("neighbour object has ", nsg, " sub-graphs")
+  }
   new.nb
  }
 
@@ -55,6 +59,10 @@ intersect.nb<-function(nb.obj1, nb.obj2){
   attr(new.nb,"type")<-paste("intersect(",attr(nb.obj1,"type"),
                              ",",attr(nb.obj2,"type"),")")
   class(new.nb)<-"nb"
+  if (get.SubgraphOption()) {
+    nsg <- n.comp.nb(new.nb)$nc
+    if (nsg > 1) warning("neighbour object has ", nsg, " sub-graphs")
+  }
   new.nb
 }
 setdiff.nb<-function(nb.obj1, nb.obj2){
@@ -99,6 +107,10 @@ setdiff.nb<-function(nb.obj1, nb.obj2){
   	attr(new.nb,"type")<-paste("setdiff(",attr(nb.obj1,"type"),
                              ",",attr(nb.obj2,"type"),")")
   	class(new.nb)<-"nb"
+        if (get.SubgraphOption()) {
+          nsg <- n.comp.nb(new.nb)$nc
+          if (nsg > 1) warning("neighbour object has ", nsg, " sub-graphs")
+        }
   	new.nb
 }
 
@@ -122,5 +134,9 @@ complement.nb<-function(nb.obj){
   }
   attr(new.nb,"type")<-paste("complement(",attr(nb.obj,"type"),")")
   class(new.nb)<-"nb"
+  if (get.SubgraphOption()) {
+    nsg <- n.comp.nb(new.nb)$nc
+    if (nsg > 1) warning("neighbour object has ", nsg, " sub-graphs")
+  }
   new.nb
  }