+++ R/read.gal.R
@@ -64,6 +64,10 @@ read.gal <- function(file, region.id=NULL, override.id=FALSE)
 	attr(res1, "gal") <- TRUE
 	attr(res1, "call") <- TRUE
 	res1 <- sym.attr.nb(res1)
+        if (get.SubgraphOption()) {
+          nsg <- n.comp.nb(res1)$nc
+          if (nsg > 1) warning("neighbour object has ", nsg, " sub-graphs")
+        }
 	res1
 }